command: let frame-step go through multiple frames
This commit gives the frame-step command the ability to accept an
additional argument specifying the amount of frames (forwards or
backwards) to step through. By default, it steps forward 1. The
frame-back-step command is reworked to simply be a call to frame-step
with a -1 value.

An additional flag is added that controls how exactly mpv steps through
frames. By default, "play" is used which behaves like before. mpv will
play through the requested amount of frames before stopping. This is
more robust and reliable but could also take much longer if a high
amount of frames is requested to be skipped through. This mode only
works if mpv is skipping frames forward. Otherwise, the "seek" flag is
used which performs a very exact seek. This working well, depends on pts
values being reliable which may not always be the case. Frame skipping
like this is more of a guess and should be less accurate.

As a minor note, MPSEEK_BACKSTEP is renamed to MPSEEK_FRAMESTEP since
forward seeks can use this as well. Fixes #10128.