ta: remove MPV_LEAK_REPORT
The ta code had some basic leak checking/reporting. This was always
mainly for development usage, but it's pretty obsolete. Nowadays with a
couple of meson flags, you can enable ASAN which is much smarter and
will catch everything (not just stuff allocated with talloc). For
platforms where LSAN is not supportered, there is superior tooling
available anyway. There's zero reason to keep supporting this which is
functionally dead code anyway, so cut the bloat.