demux/packet_pool: add fast quit mode
A new field, MPContext::quit_fast, has been introduced. Setting this
field to true bypasses the time-consuming packet deallocation process.
This is particularly useful when exiting the process, as there's no need
to iterate over and free all packets. The cache, stored as a linked list
with potentially millions of small packets, requires significant time to
iterate through and deallocate, especially if the memory is swapped out.

This change significantly speeds up the player closure, especially when
a large demuxer cache is used.

Fixes: #12294
Fixes: #12563
Signed-off-by: Kacper Michajłow <kasper93@gmail.com>