console.lua: scale with the window height
If --osd-scale-by-window is true scale console.lua with the window
height to be consistent with the OSD, osc.lua and stats.lua.

Also remove the scale option because you can set font_size and
border_size to have the same effect.

display-hidpi-scale is still used and it may make sense to also multiple
other mpv text sizes by it.