wayland: add support for frog-color-management-v1
The big color-management MR in wayland-protocols has been in development
for a very long time and honestly that protocol is super complex and
painful to parse. The frog protocol stuff is relatively new but it turns
out that kwin actually has support for this. Additionally, Arch and
Fedora even package these protocols. Given that mpv is certainly a
client that benefits from color stuff, it's not too much work to plug in
this simple protocol so users can actually start using it. This adds an
--wayland-colorspace-hint option which passes hdr metadata to the
compositor. Unlike the existing solution (using vulkan), this is
completely graphics API agnostic and will work with any wayland backend.